backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "avk-backend"
version = "0.1.0"
description = "Phase 0 backend bootstrap for the AVK platform."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Django>=5.1,<5.2",
  "djangorestframework>=3.15,<3.16",
  "psycopg[binary]>=3.2,<3.3",
  "celery>=5.4,<5.5",
  "python-dotenv>=1.0,<2.0",
  "redis>=5.0,<6.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.10,<25.0",
  "pytest>=8.3,<9.0",
  "pytest-django>=4.9,<5.0",
  "ruff>=0.7,<0.8",
]

[tool.hatch.build.targets.wheel]
packages = ["apps", "config"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]

[tool.ruff.lint.per-file-ignores]
"apps/*/migrations/*.py" = ["E501", "I001"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
pythonpath = ["."]
testpaths = ["tests", "apps"]
